refactor: LottoService 변경
LottoService가 굳이 필요하지 않다고 판단. OutputView와 InputView가 그 역할을 대신함
