+++ src/main/java/lotto/service/LottoServiceImpl.java
@@ -1,0 +1,22 @@
+package lotto.service;
+
+import java.util.List;
+import lotto.Lotto;
+
+public class LottoServiceImpl implements LottoService {
+
+    @Override
+    public List<Lotto> buyLotto(int money) {
+        return List.of();
+    }
+
+    @Override
+    public void calWinning() {
+
+    }
+
+    @Override
+    public Float revenue() {
+        return 0f;
+    }
+}
