--- src/main/java/lotto/service/ViewService.java
@@ -1,8 +1,0 @@
-package lotto.service;
-
-import java.util.List;
-
-public interface ViewService {
-    void validateMoney(int money);
-    void validateNumberSize(List<Integer> winningNumbers);
-}
