--- src/main/java/lotto/service/ViewServiceImpl.java
@@ -1,15 +1,0 @@
-package lotto.service;
-
-import java.util.List;
-
-public class ViewServiceImpl implements ViewService {
-    @Override
-    public void validateMoney(int money) {
-
-    }
-
-    @Override
-    public void validateNumberSize(List<Integer> winningNumbers) {
-
-    }
-}
