+++ src/test/java/lotto/controller/ViewControllerTest.java
@@ -1,17 +1,14 @@
-package lotto.service;
+package lotto.controller;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.*;
 
-import java.util.ArrayList;
 import java.util.List;
-import lotto.Lotto;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-class ViewServiceTest {
-
-    ViewService viewService;
+class ViewControllerTest {
+    ViewController viewController;
 
     @BeforeEach
     void setUp() {
@@ -20,14 +17,15 @@ class ViewServiceTest {
 
     @Test
     void validateMoney() {
-        assertThatThrownBy(() -> viewService.validateMoney(100))
+        assertThatThrownBy(() -> viewController.validateMoney(100))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 
     @Test
     void validateNumberSize() {
         List<Integer> winningNumbers = List.of(1,2,3,4,5,6,7);
-        assertThatThrownBy(()-> viewService.validateNumberSize(winningNumbers))
+        assertThatThrownBy(()-> viewController.validateNumberSize(winningNumbers))
                 .isInstanceOf(IllegalArgumentException.class);
     }
+
 }
